improved import progress

+++ fragmentation_tree/fragmentation_tree_construction/src/main/java/de/unijena/bioinf/FragmentationTreeConstruction/computation/FasterTreeComputationInstance.java
@@ -123,7 +123,7 @@ public class FasterTreeComputationInstance extends BasicMasterJJob<FasterTreeCom
     }
 
     protected void configureProgress(int to, int numberOfTicks) {
-        configureProgress(Math.min(to-1,this.currentProgress().getNewValue()), to, numberOfTicks);
+        configureProgress(Math.min(to-1, this.currentProgress().getNewValue().intValue()), to, numberOfTicks);
     }
 
     protected void configureProgress(int from, int to, int numberOfTicks) {
