+++ gibbs_sampling/src/main/java/de/unijena/bioinf/GibbsSampling/model/GraphBuilder.java
@@ -285,10 +285,10 @@ public class GraphBuilder<C extends Candidate<?>> extends BasicMasterJJob<Graph<
 
     @Override
     public void progressChanged(JobProgressEvent progressEvent) {
-        if (progressEvent.getNewValue()!=100) return;
+        if (progressEvent.getNewValue().longValue() !=100) return;
         ++numberOfFinishedComputations;
         if(numberOfFinishedComputations % step == 0 || numberOfFinishedComputations==size) {
-            logInfo(Math.round(100*(numberOfFinishedComputations)/size)+"%");
+            logInfo(Math.round(100*((double)numberOfFinishedComputations)/size)+"%");
         }
     }
